# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RAKT KAVACH - National Digital Blood Grid API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory:
    get:
      operationId: listInventory
      tags: [inventory]
      summary: List blood inventory
      parameters:
        - name: bloodGroup
          in: query
          schema:
            type: string
        - name: bloodBankId
          in: query
          schema:
            type: integer
        - name: hospitalId
          in: query
          schema:
            type: integer
        - name: state
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Inventory list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryListResponse"
...
  /inventory/search:
    get:
      operationId: searchInventory
      tags: [inventory]
      summary: Search available blood
      parameters:
        - name: bloodGroup
          in: query
          required: true
          schema:
            type: string
        - name: units
          in: query
          schema:
            type: integer
        - name: state
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventorySearchResponse"

  # BLOOD UNITS
  /blood-units:
    get:
      operationId: listBloodUnits
      tags: [blood-units]
      summary: List blood units
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: bloodGroup
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Blood units list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodUnitListResponse"
    post:
      operationId: createBloodUnit
      tags: [blood-units]
      summary: Register a new blood unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodUnitInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodUnit"
  /blood-units/{id}:
    get:
      operationId: getBloodUnit
      tags: [blood-units]
      summary: Get blood unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blood unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodUnit"
    patch:
      operationId: updateBloodUnit
      tags: [blood-units]
      summary: Update blood unit status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodUnitUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodUnit"
  /blood-units/{id}/timeline:
    get:
      operationId: getBloodUnitTimeline
      tags: [blood-units]
      summary: Get blood unit lifecycle timeline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodUnitTimelineResponse"

  # BLOOD REQUESTS
  /blood-requests:
    get:
      operationId: listBloodRequests
      tags: [inventory]
      summary: List blood requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: bloodGroup
          in: query
          schema:
            type: string
        - name: hospitalId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Blood requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequestListResponse"
    post:
      operationId: createBloodRequest
      tags: [inventory]
      summary: Create blood request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodRequestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"
  /blood-requests/{id}:
    get:
      operationId: getBloodRequest
      tags: [inventory]
      summary: Get blood request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blood request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"
    patch:
      operationId: updateBloodRequest
      tags: [inventory]
      summary: Update blood request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodRequestUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"

  # EMERGENCY SOS
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # AUTH SCHEMAS
...
    Inventory:
      type: object
      required: [id, bloodBankId, bloodGroup, units, collectionDate, expiryDate, status, createdAt]
      properties:
        id:
          type: integer
        bloodBankId:
          type: integer
        bloodBankName:
          type: string
        bloodGroup:
          type: string
        units:
          type: number
        collectionDate:
          type: string
        expiryDate:
          type: string
        status:
          type: string
          enum: [AVAILABLE, RESERVED, DISPATCHED, EXPIRED, DISCARDED]
        donorId:
          type: ["integer", "null"]
        createdAt:
          type: string
    InventoryListResponse:
      type: object
      required: [inventory, total, page, limit]
      properties:
        inventory:
          type: array
          items:
            $ref: "#/components/schemas/Inventory"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    InventorySearchResult:
      type: object
      required: [bloodBankId, bloodBankName, bloodGroup, units, state, district, distance]
      properties:
        bloodBankId:
          type: integer
        bloodBankName:
          type: string
        bloodGroup:
          type: string
        units:
          type: number
        state:
          type: string
        district:
          type: string
        distance:
          type: ["number", "null"]
        contactNumber:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
    InventorySearchResponse:
      type: object
      required: [results, totalFound, bloodGroup]
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/InventorySearchResult"
        totalFound:
          type: integer
        bloodGroup:
          type: string

    # BLOOD UNIT TRACKING
    BloodUnitInput:
      type: object
      required: [bloodGroup, donorId, collectionDate]
      properties:
        bloodGroup:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        donorId:
          type: integer
        collectionDate:
          type: string
        bloodBankId:
          type: integer
    BloodUnitUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [DONATED, COLLECTED, TESTING, LAB_VALIDATION, STORAGE, RESERVED, DISPATCHED, IN_TRANSIT, AT_HOSPITAL, TRANSFUSED, EXPIRED, DISCARDED]
        currentLocation:
          type: string
        hospitalId:
          type: ["integer", "null"]
        notes:
          type: string
    BloodUnit:
      type: object
      required: [id, unitCode, bloodGroup, donorId, collectionDate, expiryDate, status, createdAt]
      properties:
        id:
          type: integer
        unitCode:
          type: string
        bloodGroup:
          type: string
        donorId:
          type: integer
        donorName:
          type: ["string", "null"]
        collectionDate:
          type: string
        expiryDate:
          type: string
        status:
          type: string
        currentLocation:
          type: ["string", "null"]
        bloodBankId:
          type: ["integer", "null"]
        bloodBankName:
          type: ["string", "null"]
        hospitalId:
          type: ["integer", "null"]
        hospitalName:
          type: ["string", "null"]
        createdAt:
          type: string
    BloodUnitListResponse:
      type: object
      required: [bloodUnits, total, page, limit]
      properties:
        bloodUnits:
          type: array
          items:
            $ref: "#/components/schemas/BloodUnit"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    TimelineEvent:
      type: object
      required: [status, timestamp, location]
      properties:
        status:
          type: string
        timestamp:
          type: string
        location:
          type: string
        notes:
          type: ["string", "null"]
        updatedBy:
          type: ["string", "null"]
    BloodUnitTimelineResponse:
      type: object
      required: [unitId, unitCode, bloodGroup, timeline]
      properties:
        unitId:
          type: integer
        unitCode:
          type: string
        bloodGroup:
          type: string
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"

    # BLOOD REQUEST SCHEMAS
    BloodRequestInput:
      type: object
      required: [bloodGroup, units, urgency, patientId, hospitalId]
      properties:
        bloodGroup:
          type: string
          enum: [A+, A-, B+, B-, AB+, AB-, O+, O-]
        units:
          type: number
        urgency:
          type: string
          enum: [ROUTINE, URGENT, EMERGENCY]
        patientId:
          type: integer
        hospitalId:
          type: integer
        requiredBy:
          type: string
        notes:
          type: string
    BloodRequestUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [PENDING, APPROVED, FULFILLED, PARTIALLY_FULFILLED, CANCELLED, REJECTED]
        bloodBankId:
          type: integer
        notes:
          type: string
    BloodRequest:
      type: object
      required: [id, bloodGroup, units, urgency, status, createdAt]
      properties:
        id:
          type: integer
        bloodGroup:
          type: string
        units:
          type: number
        urgency:
          type: string
        status:
          type: string
        patientId:
          type: integer
        patientName:
          type: ["string", "null"]
        hospitalId:
          type: integer
        hospitalName:
          type: ["string", "null"]
        bloodBankId:
          type: ["integer", "null"]
        bloodBankName:
          type: ["string", "null"]
        requiredBy:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    BloodRequestListResponse:
      type: object
      required: [requests, total, page, limit]
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/BloodRequest"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # EMERGENCY SCHEMAS
...
    Notification:
      type: object
      required: [id, userId, type, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum: [EMERGENCY_SOS, DONATION_REMINDER, ELIGIBILITY_RESTORED, BLOOD_AVAILABLE, REQUEST_FULFILLED, IMPACT_UPDATE, SYSTEM, REWARD_EARNED]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        actionUrl:
          type: ["string", "null"]
        metadata:
          type: ["string", "null"]
        createdAt:
          type: string
    NotificationListResponse:
      type: object
      required: [notifications, total, unreadCount, page, limit]
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total:
          type: integer
        unreadCount:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ANALYTICS SCHEMAS
...